Support 64 bit integers in JSON models (#113)
* Allow JSON numbers larger than 32 bits

* Update readme

+++ Engine/Model/Helpers/JsonGraph.cs
@@ -54,7 +54,7 @@ namespace Engine.Model.Helpers
                     return p.Value<string>();
 
                 case JTokenType.Integer:
-                    return p.Value<int>();
+                    return p.Value<long>();
 
                 case JTokenType.Float:
                     return p.Value<double>();
